requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Deep learning
torch==2.0.1
sentence-transformers==2.2.2

# Vector database
chromadb==0.4.13
onnxruntime==1.15.1

# Graph processing
networkx==3.1
pyvis==0.3.2

# NLP and text processing
transformers==4.30.2
cross-encoder==0.2.2

# API integration
openai==1.3.5
aiohttp==3.8.5
async-timeout==4.0.3

# Utilities
tqdm==4.65.0
pydantic==1.10.12
python-dotenv==1.0.0